merge-edgelists -> add --rm-edgelists cli arg; do not throw error on executing rm -rf ..; mv existing *.edgelist.csv file to data.edgelist.csv

+++ lodcc_merge_edgelists.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 FOLDER_ROOT=${1:-"dumps/"}
-USE_CACHE=${2:-true}
+RM_EDGELISTS=${2:-False}
 
 # remove backslash from FOLDER if present
 if [[ "${FOLDER_ROOT%/}" != "$FOLDER_ROOT" ]]; then
@@ -12,10 +12,16 @@ FILES_COUNT=`find "$FOLDER_ROOT" -name "*.edgelist.csv" -type f | wc -l`
 
 # removedata.edgelist, if present
 if [[ $FILES_COUNT > 1 ]]; then
-    rm "$FOLDER_ROOT/data.edgelist.csv"
+    rm "$FOLDER_ROOT/data.edgelist.csv" &> /dev/null # ignore errors when file does not exist, for instance
 else
     # if data.edgelist is the only file, we're fine
     if [ -f "$FOLDER_ROOT/data.edgelist.csv" ]; then
+        exit 0
+    else
+        # else take that one edgelist file and rename it
+        FILE=`ls $FOLDER_ROOT/*.edgelist.csv`
+        FOLDER=${FILE%/*}
+        mv "$FILE" "$FOLDER/data.edgelist.csv"
         exit 0
     fi
 fi
@@ -27,8 +33,11 @@ for file in $FILES; do
     # ignore data.edgelist, if present
     if [[ "${file%*/data.edgelist.csv}" == "$file" ]]; then
         FOLDER=${file%/*}
-        cat $file >> "$FOLDER/data.edgelist.csv" &&  
-        rm $file
+        cat $file >> "$FOLDER/data.edgelist.csv"
+        
+        if [[ $RM_EDGELISTS = "True" ]]; then
+            rm $file
+        fi
     fi
 
 done
